ditch the old mdoutline, based on marked AST.
